only installed required depedencies in each benchmark env
