+++ tests/benchmarks/test_read_tensorstore_benchmark.py
@@ -8,8 +8,8 @@ from tests.benchmarks.benchmark_parameters import (
     GZIP_LEVEL,
     ZSTD_LEVEL,
 )
-from zarr_benchmarks import read_write_tensorstore
 
+read_write_tensorstore = pytest.importorskip("zarr_benchmarks.read_write_tensorstore")
 pytestmark = [pytest.mark.tensorstore]
