+++ tests/benchmarks/test_write_tensorstore_benchmark.py
@@ -8,9 +8,9 @@ from tests.benchmarks.benchmark_parameters import (
     GZIP_LEVEL,
     ZSTD_LEVEL,
 )
-from zarr_benchmarks import read_write_tensorstore
 from zarr_benchmarks.utils import remove_output_dir
 
+read_write_tensorstore = pytest.importorskip("zarr_benchmarks.read_write_tensorstore")
 pytestmark = [pytest.mark.tensorstore]
