+++ tests/benchmarks/test_write_zarr_python_benchmark.py
@@ -13,9 +13,7 @@ from zarr_benchmarks.utils import remove_output_dir
 try:
     from zarr_benchmarks import read_write_zarr_v3 as read_write_zarr
 except ImportError:
-    from zarr_benchmarks import (  # type: ignore[no-redef]
-        read_write_zarr_v2 as read_write_zarr,
-    )
+    read_write_zarr = pytest.importorskip("zarr_benchmarks.read_write_zarr_v2")
 
 pytestmark = [pytest.mark.zarr_python]
