+++ tests/tests/test_write_zarr_array.py
@@ -1,51 +1,17 @@
-from pathlib import Path
-from typing import Callable
-
-import numpy as np
 import pytest
 
-from zarr_benchmarks import read_write_tensorstore
+import tests.tests.utils as utils
 
 try:
     from zarr_benchmarks import read_write_zarr_v3 as read_write_zarr
 except ImportError:
-    from zarr_benchmarks import (  # type: ignore[no-redef]
-        read_write_zarr_v2 as read_write_zarr,
-    )
+    read_write_zarr = pytest.importorskip("zarr_benchmarks.read_write_zarr_v2")
+
+pytestmark = [pytest.mark.zarr_python]
 
 
-def assert_empty_chunks_written(
-    output_dir: Path, write_zarr_array: Callable, write_empty_chunks: bool
-) -> None:
-    """Check an empty chunk is written to file when write_empty_chunks=True"""
-
-    image = np.zeros(shape=(1, 1, 1))
-    store_path = output_dir / "image.zarr"
-
-    write_zarr_array(
-        image,
-        store_path,
-        overwrite=True,
-        chunks=(1, 1, 1),
-        compressor=None,
-        write_empty_chunks=write_empty_chunks,
-    )
-
-    assert (store_path / ".zarray").exists()
-    assert (store_path / "0.0.0").exists() == write_empty_chunks
-
-
-@pytest.mark.zarr_python
 @pytest.mark.parametrize("write_empty_chunks", [True, False])
 def test_zarr_python_write_empty_chunks(tmp_path, write_empty_chunks):
-    assert_empty_chunks_written(
+    utils.assert_empty_chunks_written(
         tmp_path, read_write_zarr.write_zarr_array, write_empty_chunks
-    )
-
-
-@pytest.mark.tensorstore
-@pytest.mark.parametrize("write_empty_chunks", [True, False])
-def test_tensorstore_write_empty_chunks(tmp_path, write_empty_chunks):
-    assert_empty_chunks_written(
-        tmp_path, read_write_tensorstore.write_zarr_array, write_empty_chunks
     )
