+++ tox.toml
@@ -12,8 +12,8 @@ commands = [
         {extend = true, replace = "posargs"},
     ],
 ]
-deps = ["tensorstore==0.1.73"]
 description = "Run benchmarks under tensorstore"
+extras = ["tensorstore"]
 
 [env.py313-zarrv2]
 commands = [
@@ -26,8 +26,8 @@ commands = [
         {extend = true, replace = "posargs"},
     ],
 ]
-deps = ["zarr==2.18.4", "numcodecs==0.15.1"]
 description = "Run benchmarks under zarr-python version 2"
+extras = ["zarr-python-v2"]
 
 [env.py313-zarrv3]
 commands = [
@@ -40,5 +40,5 @@ commands = [
         {extend = true, replace = "posargs"},
     ],
 ]
-deps = ["zarr==3.0.6", "numcodecs==0.15.1"]
 description = "Run benchmarks under zarr-python version 3"
+extras = ["zarr-python-v3"]
